Added language controller class
Language controller that processes a Time object and allows the english
spoken string to be returned using methods to work out the tense of the
sentence or whether to include the minutes if it is on the hour (e.g
10:00) exactly.

Added getter methods to the Time object. Added to the testbed class to
test the LanguageController class. Need to add functionality to convert
the time digits to english words using array or external dictionary
file.